+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.27 (2021-08-24)
+
+**Note:** Version bump only for package @winfe/winex-cli
+
+
+
+
+
+## 1.1.26 (2021-08-23)
+
+**Note:** Version bump only for package @winfe/winex-cli
+
+
+
+
+
+## 1.1.25 (2021-08-23)
+
+**Note:** Version bump only for package @winfe/winex-cli
+
+
+
+
+
 ## 1.1.24 (2021-08-20)